requirements.txt
flask
python-dotenv
google-generativeai
groq
pdfplumber
pytesseract
Pillow
pymongo
pdfkit>=1.0.0

weasyprint>=60.0  # PDF generation (pure Python, no Node.js)
